Xen Virtual Appliance Auto Upgrade Script
A script for automatically upgrading a xenserver vm. This requires all
the upgradable software to reside on a dedicated virtual disk attached
to the vm. So the script simply build a new disk using new software and
detaches the disk containing the old software and attaches a new disk
with the upgraded software.